support compiling outside git checkout
When building nmrpflash for Debian, there is no 'git' command available,
and there is no '.git' directory either. This makes the build emit
warnings from the $(shell) calls in VERSION variable in the Makefile
and breaks the '-V' option.

The change in this patch accounts for missing 'git' command and resorts
to using a STANDALONE_VERSION environment variable which in turn is
provided by the package build mechanics.

This change has no effect when the git command is available and the
'.git' directory is present.